FeaturePreferences: Introduce settings reset
This introduces a new selectable settings reset flow under
"Backup and reset." The user is presented with a list of
settings that were set via OTS and given the choice to reset
a select few of them or reset all of them. This will cause OTS
to show the setting choice snackbar again next time the feature
is used.

Change-Id: Ic7e787834d3bb01b2a6ce0a01ecaaf1a85b64d34
Signed-off-by: Evan Anderson <evananderson@aospa.co>

FeaturePreferences: Revise layout dimensions

These new dimensions make the layout compatible with RTL
languages and fix padding to match Arz's specification.

Change-Id: I0fd631bc51ced210401f9d29601590f3f6a2971d
Signed-off-by: Evan Anderson <evananderson@aospa.co>

FeaturePreferences: Add setting summary

This adds a short setting summary to the privacy settings screen.

Change-Id: Iab533596623174ee2e763b2a07122ddd9d7b6031
Signed-off-by: Evan Anderson <evananderson@aospa.co>

FeaturePreferences: Major refactor of settings handling

This is a big change that replaces the custom view mess with an implementation
of CheckBoxPreference inside of a ListView. This change also includes code and
layout cleanup. It should not change the appearance of the view.

Change-Id: I53b0d193ef6893f8ee12e1e95839fc085a5c92bd
Signed-off-by: Evan Anderson <evananderson@aospa.co>

Feature Preferences: Don't hide settings that aren't set

Instead of completely hiding the OTS settings that haven't been set
or that have been reset, just disable them so the user can still see
that they are there. Change suggested by Arz.

This commit also includes minor code cleanup

Change-Id: I6932ea007323a3054b23c7284e0970ac622a3b6a
Signed-off-by: Evan Anderson <evananderson@aospa.co>
Signed-off-by: klozz <shaclone707@gmail.com>