Settings: Expose feature flag UI on all builds
The feature flag controller UI is already hidden in developer options,
and the underlying functionality is still present in non-debuggable
builds — it's just harder to access.

Expose the feature flag controller UI on all build types to make it
easier for users to change them, e.g. to enable/disable FUSE.

Change-Id: I027eace8ea80177b9a301407bbfab3e44e262954