Wifi: do not throw exception when wifi res context not found
When prebuilt mainline modules is enabled, the package
wifi.resources" is overriden build time with
other wifi.resources, causing crash as wifi res is invalid.
In case res context is not found set mWifiRes and
mWifiResContext to null, to avoid crash when
wifi.resources is not used, as mainline module is enabled.

CRs-Fixed: 3378048
Change-Id: I79c59a6c0b603242f29b6fc276fdcc7c542bd6ec